+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+Future version
+
+- Added support for bigBed files
+
 v1.13.0
 
 - Add support from cooler v2 files for the fragments API